Rename AIV4 and AIV5
