+++ README.md
@@ -4,4 +4,4 @@ This is an interactive, text-based version of the card game Euchre. https://bicy
 The human user is player 1, on a team with player 3 against the other 2 plays. Cards used are 9 -> Ace ascending.
 The goal is to win more tricks than the other team. The first team to 11 points win.
 
-To play the game, type `python Euchre.py`.
+To play the game, type `python main.py`. To change which AI version is playing, edit the `config.json` file with the appropriate version name and edit line 15 in `main.py` to be testing = False.
